tests: Improve and re-organize tests and CI (#57)
