+++ tests/conftest.py
@@ -2,6 +2,9 @@
 
 from __future__ import annotations
 
+import os
+import warnings
+
 from pathlib import Path
 
 import pytest
@@ -13,6 +16,9 @@ from dotenv import dotenv_values, load_dotenv
 def env() -> dict[str, str]:
     """Return a fixture for the environment variables from .env."""
     dotenv_file = Path(__file__).parent / '.env'
+    if not dotenv_file.exists():
+        warnings.warn('No .env file found.')
+        return {}
     load_dotenv(dotenv_file)
     return dotenv_values(dotenv_file)
 
@@ -24,3 +30,36 @@ def animals_data() -> list[str]:
     with open(data_path) as f:
         data = [line.strip() for line in f.readlines() if line.strip()]
         return data
+
+
+@pytest.fixture
+def api_key_openai(env: dict[str, str]) -> str:
+    """Fixture for OpenAI API key from environment."""
+    api_key = os.getenv('OPENAI_API_KEY')
+    if not api_key:
+        raise EnvironmentError(
+            'Please set the OPENAI_API_KEY environment variable.'
+        )
+    return api_key
+
+
+@pytest.fixture
+def api_key_gemini(env: dict[str, str]) -> str:
+    """Fixture for Gemini API key from environment."""
+    api_key = os.getenv('GEMINI_API_KEY')
+    if not api_key:
+        raise EnvironmentError(
+            'Please set the GEMINI_API_KEY environment variable.'
+        )
+    return api_key
+
+
+@pytest.fixture
+def api_key_hugging_face(env: dict[str, str]) -> str:
+    """Fixture for Hugging Face API key from environment."""
+    api_key = os.getenv('HF_TOKEN')
+    if not api_key:
+        raise EnvironmentError(
+            'Please set the GEMINI_API_KEY environment variable.'
+        )
+    return api_key
